Add withoutEncodingUrl flag
Both ACF and Lucee encode the url passed to CFHTTP.  This is less
than desirable since it encodes valid url characters such as a
colon (`:`).  In Lucee, we can skip this automatic encoding using the
`encodeurl` flag.  There is no such flag in ACF.  This change adds a
flag in the `HyperRequest` called `encodeUrl` which is set to `true`
by default.  If it is `true` when sending the request, nothing will
be added to the attribute struct.  If it is `false`, it will be
added to the attribute struct passed to `cfhttp`.  This will cause
an error in ACF, but there's really not much of a workaround right now.