feat: Allow adding multiple query params with the same key.
On the backs of 2cb963e4f761479f1cfad06b8949ce00d2170a7d, this commit documents
the methods to use to `append` new query params without overwriting
the existing values with the same key.

BREAKING CHANGE: The `getQueryParams` and `setQueryParams` methods have
changed to represent the new internal structure of query params in Hyper.
These now are represented internally as an array of structs to support
duplicate keys in query strings.  You _probably_ want the
`getQueryParamByName` and `getAllQueryParamsByName` instead of
`getQueryParams` and `withQueryParams` or `appendQueryParams` methods
instead of `setQueryParams`. One more item of note, the `getQueryParam`\
method has been deprecated in favor of `getQueryParamByName`.