feat(HyperRequest): Add retry handling to HyperRequests
HyperRequests can now be set to retry requests.  You can provide
either a number of retries and a retry delay value (in milliseconds)
or an array of retry delay values (to account for exponential backoffs).
Additionally, a predicate function can be provided to determine if a
request should be retried and to even modify the next request to be sent.
(The default predicate function is `return HyperResponse.isError();`.)